Flask/requirements.txt
opencv-python==4.10.0.84
numpy==2.1.2
Flask==3.0.3
paho-mqtt==2.1.0
imgbeddings==0.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
huggingface-hub==0.25.1
flask-socketio==5.4.1